+++ app/src/main/java/com/rdc/androidtestappa/MainActivity.java
@@ -28,6 +28,7 @@ import java.time.Month;
 import java.util.ArrayList;
 import java.util.Date;
 
+
 public class MainActivity extends AppCompatActivity {
 
     /**
@@ -67,23 +68,22 @@ public class MainActivity extends AppCompatActivity {
     }
 
 
-    @Override
-    public boolean onOptionsItemSelected(MenuItem item) {
-        // Handle action bar item clicks here. The action bar will
-        // automatically handle clicks on the Home/Up button, so long
-        // as you specify a parent activity in AndroidManifest.xml.
-        int id = item.getItemId();
-
-        //noinspection SimplifiableIfStatement
-        if (id == R.id.action_sortByDate) {
-
-            return true;
-        }
-        if (id == R.id.action_sortByStatus) {
-
-            return true;
-        }
-        return super.onOptionsItemSelected(item);
-    }
+//    @Override
+//    public boolean onOptionsItemSelected(MenuItem item) {
+//        // Handle action bar item clicks here. The action bar will
+//        // automatically handle clicks on the Home/Up button, so long
+//        // as you specify a parent activity in AndroidManifest.xml.
+//        int id = item.getItemId();
+//
+//        //noinspection SimplifiableIfStatement
+//        if (id == R.id.action_sortByDate) {
+//            return true;
+//        }
+//        if (id == R.id.action_sortByStatus) {
+//
+//            return true;
+//        }
+//        return super.onOptionsItemSelected(item);
+//    }
 
 }
